+++ zerofc.com/static.config.js
@@ -2,11 +2,14 @@ import path from 'path'
 import axios from 'axios'
 import React from 'react'
 
+import Document from './src/document'
+
 const SITE_ROOT = 'https://zerofc.web.fc2.com'
 const SITE_TITLE = 'ゼロ戦クラブ'
 const SITE_SUB_TITLE = '岡山総合格闘技ジム'
 
 export default {
+  Document,
   siteRoot: SITE_ROOT,
   getSiteData: async ({ dev }) => ({
     title: SITE_TITLE,
@@ -136,41 +139,6 @@ export default {
         })
       }
     ]
-  },
-  Document: ({
-    Html,
-    Head,
-    Body,
-    children,
-    state: { siteData, renderMeta },
-  }) => {
-    return (
-      <Html lang="ja">
-        <Head>
-          <meta charSet="UTF-8" />
-          <meta name="viewport" content="width=device-width, initial-scale=1" />
-          <meta name="google-site-verification" content="VY6apQgrtzLTC_yR2i9QmKyTyc4xV9Injys1efZocMU" />
-          <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/solid.css" integrity="sha384-QokYePQSOwpBDuhlHOsX0ymF6R/vLk/UQVz3WHa6wygxI5oGTmDTv8wahFOSspdm" crossOrigin="anonymous" />
-          <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/fontawesome.css" integrity="sha384-vd1e11sR28tEK9YANUtpIOdjGW14pS87bUBuOIoBILVWLFnS+MCX9T6MMf0VdPGq" crossOrigin="anonymous" />
-          <script
-            dangerouslySetInnerHTML={{
-              __html: `
-                var _gaq = _gaq || [];
-                _gaq.push(['_setAccount', 'UA-2987422-2']);
-                _gaq.push(['_trackPageview']);
-
-                (function() {
-                    var ga = document.createElement('script'); ga.type = 'text/javascript'; ga.async = true;
-                    ga.src = ('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';
-                    var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(ga, s);
-                })();
-              `,
-            }}
-          />
-        </Head>
-        <Body>{children}</Body>
-      </Html>
-    )
   },
   plugins: [
     [
